move scripts to simple directory
