+++ simple/utils/simple_utils.py
@@ -1,0 +1,87 @@
+"""
+Utils functions for use in ingests
+"""
+
+import logging
+from astroquery.simbad import Simbad
+from astropy.table import Table
+
+
+__all__ = [
+    "find_survey_name_in_simbad",
+]
+
+logger = logging.getLogger("SIMPLE")
+
+
+# SURVEY DATA
+def find_survey_name_in_simbad(sources, desig_prefix, source_id_index=None):
+    """
+    Function to extract source designations from SIMBAD
+
+    Parameters
+    ----------
+    sources: astropy.table.Table
+        Sources names to search for in SIMBAD
+    desig_prefix
+        prefix to search for in list of identifiers
+    source_id_index
+        After a designation is split, this index indicates source id suffix.
+        For example, source_id_index = 2 to extract suffix from "Gaia DR2" designations.
+        source_id_index = 1 to exctract suffix from "2MASS" designations.
+    Returns
+    -------
+    Astropy table
+    """
+
+    n_sources = len(sources)
+
+    Simbad.reset_votable_fields()
+    Simbad.add_votable_fields("typed_id")  # keep search term in result table
+    Simbad.add_votable_fields("ids")  # add all SIMBAD identifiers as an output column
+
+    logger.info("simbad query started")
+    result_table = Simbad.query_objects(sources["source"])
+    logger.info("simbad query ended")
+
+    ind = result_table["SCRIPT_NUMBER_ID"] > 0  # find indexes which contain results
+    simbad_ids = result_table["TYPED_ID", "IDS"][ind]
+
+    db_names = []
+    simbad_designations = []
+    source_ids = []
+
+    for row in simbad_ids:
+        db_name = row["TYPED_ID"]
+        ids = row["IDS"].split("|")
+        designation = [i for i in ids if desig_prefix in i]
+
+        if designation:
+            logger.debug(f"{db_name}, {designation[0]}")
+            db_names.append(db_name)
+            if len(designation) == 1:
+                simbad_designations.append(designation[0])
+            else:
+                simbad_designations.append(designation[0])
+                logger.warning(f"more than one designation matched, {designation}")
+
+            if source_id_index is not None:
+                source_id = designation[0].split()[source_id_index]
+                source_ids.append(int(source_id))  # convert to int since long in Gaia
+
+    n_matches = len(db_names)
+    logger.info(
+        f"Found, {n_matches}, {desig_prefix}, sources for, {n_sources}, sources"
+    )
+
+    if source_id_index is not None:
+        result_table = Table(
+            [db_names, simbad_designations, source_ids],
+            names=("db_names", "designation", "source_id"),
+        )
+    else:
+        result_table = Table(
+            [db_names, simbad_designations], names=("db_names", "designation")
+        )
+
+    return result_table
