+++ tests/test_integrity.py
@@ -3,7 +3,7 @@ import os
 import pytest
 from . import REFERENCE_TABLES
 from sqlalchemy import func, and_  # , select, except_
-from schema.schema import *
+from simple.schema import *
 from astrodbkit2.astrodb import create_database, Database, or_
 from astropy.table import unique
 from astropy import units as u
